Created a downloading interface from OpenSubtitles.
Firstly, creating a payload for the proxy to send through the OpenSubtitles API the subtitle
file information is collected which is then saved to a file. The data from
the file is then read and using aiohttp the subtitle file is asynchronously
downloaded and saved to the directory of the movie.