In subtitles.py the  _perform_query  function now stores the information gathered with xmlrpc from OpenSubtitles to a new table in  media.db called "sub_dl_links". The function  _download_and_save_file  reads those entries from the table and downloads the subtitles asynchronously.
The problem that needs to be resolved is that on line 85 in subtitles.py
the SearchSubtitles is not a coroutine. It should be somehow converted
to a coroutine so the subtitle information look up process can also be
asynchronous.

db_interactor.py has been altered to add the new table "sub_dl_links" and
a method for adding elements to that table.