Downloading from OpenSubtitles completed.
SubtitleDownloader class can call its method download_from_opensubtitles
which will log in to OpenSubtitles server (currently with TemporaryAgent)
and start querying the server for subtitle information in the method
_perform_query_and_store. The subtitle information is stored into a
table "search_subs" in a database.
After that, the method _download_file starts downloading the file via the
API. It calls proxy.DownloadSubtitles() and receives tuples of:
    (subtitle_id, byte_data)
That data is then stored to the table "download_subs" in the database.
Afterwards, that data from "download_subs" is pulled and with the name
of the subtitle file and absolute path of the movie which is retrieved
from "search_subs" and passed to _write_file which writes the file in
an unarchived format.